requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-storage>=2.0.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
timeout-decorator==0.5.0